requirements.txt
pandas
numpy
matplotlib
statsmodels
pytest
python-dotenv
doit
jupyter
